fix: update packages and content config path
