+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "joshc.uk",
   "type": "module",
-  "version": "2.1.1",
+  "version": "2.2.0",
   "scripts": {
     "dev": "astro dev",
     "start": "astro dev",
@@ -11,7 +11,7 @@
     "lint": "npx prettier . --write"
   },
   "dependencies": {
-    "@astrojs/check": "^0.9.8",
+    "@astrojs/check": "^0.9.2",
     "@astrojs/mdx": "^5.0.2",
     "@astrojs/react": "^5.0.1",
     "@astrojs/rss": "^4.0.17",
@@ -24,17 +24,21 @@
     "@types/react-dom": "^18.3.0",
     "astro": "^6.0.6",
     "astro-icon": "^1.1.1",
-    "astro-mermaid": "^1.2.0",
+    "astro-mermaid": "^2.0.1",
     "axios": "^1.13.5",
     "mermaid": "^11.12.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "tailwindcss": "^4.1.18",
-    "typescript": "^5.4.5"
+    "typescript": "^5.4.5",
+    "yaml": "^2.8.2"
   },
   "devDependencies": {
     "@iconify-icon/react": "^2.1.0",
     "prettier": "^3.3.3",
     "prettier-plugin-astro": "^0.14.1"
+  },
+    "overrides": {
+    "vite": "^7"
   }
 }
