+++ src/consts.ts
@@ -2,4 +2,4 @@
 // You can import this data from anywhere in your site by using the `import` keyword.
 
 export const SITE_TITLE = "Josh Cooper";
-export const SITE_DESCRIPTION = "Making the computers do the funnies.";
+export const SITE_DESCRIPTION = "Software Engineer working in DevOps and Robotics";
