--- src/content/config.ts
@@ -1,33 +1,0 @@
-import { defineCollection, z } from "astro:content";
-
-const blog = defineCollection({
-  type: "content",
-  // Type-check frontmatter using a schema
-  schema: z.object({
-    title: z.string(),
-    description: z.string(),
-    pubDate: z.coerce.date(),
-    update: z.coerce.date().optional(),
-    icon: z.string().optional(),
-    headerImage: z.string().optional(),
-    draft: z.boolean().default(false),
-    tags: z.array(z.string()).default([]),
-  }),
-});
-
-const portfolio = defineCollection({
-  type: "content",
-  // Type-check frontmatter using a schema
-  schema: z.object({
-    title: z.string(),
-    description: z.string(),
-    pubDate: z.coerce.date(),
-    update: z.coerce.date().optional(),
-    icon: z.string().optional(),
-    headerImage: z.string().optional(),
-    draft: z.boolean().default(false),
-    tags: z.array(z.string()).optional(),
-  }),
-});
-
-export const collections = { blog, portfolio };
